requirements.txt
yarl
multidict
mc-pyority-queue
flask
requests
mapcoloniesjsonlogger==1.0.0
boto3
gdal2tiles